Feat/mb abstract models services (#97)
* made an abstract class for news, posts, events

* add another abstract class TranslationContent for news, posts, events, blogs

* write migration solving problem in news_translation and olymp_organizations

+++ src/main/java/kg/edu/mathbilim/dto/organization/OlympOrganizationDto.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class OlympOrganizationDto {
-    private Integer olympiadId;
+    private Long olympiadId;
     private Long organizationId;
 }
