+++ src/main/java/kg/edu/mathbilim/repository/olympiad/OlympiadRepository.java
@@ -3,5 +3,5 @@ package kg.edu.mathbilim.repository.olympiad;
 import kg.edu.mathbilim.model.olympiad.Olympiad;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OlympiadRepository extends JpaRepository<Olympiad, Integer> {
+public interface OlympiadRepository extends JpaRepository<Olympiad, Long> {
 }
